Major Updates (Client and Server)
- Client.Functions.OPT.cs method optExists reduced to single line
- Client.Functions.UpdateHandler.cs direct manipulation of
GUI.Instance.totalProgress values now properly delegated through
GUI.Instance.UpdateProgressMaximum
- Client.Functions.UpdateHandler.cs checking status length on
Core.ProgressChanged event removed
- Client.Functions.UpdateHandler.cs DownloadSelfUpdate method removed
(as part of self updating overhaul)
- Client.Functions.UpdateHandler.cs OnFileTransferred event method
restructured to process data.xxx and resources updates side by side
instead of in-line. New status messages added accordingly
- Client.GUI.cs several task triggers moved and set to specific events
triggered by the Portal Server for synchronicity and to eliminate rouge
method calls causing progressbar value out of bounds resulting in crash
- Client.GUI.cs registerClient method added to inform Portal Server that
it should statistically record the Launcher as a client not updater
(statistics purposes only)
- Client.GUI.cs OnRegisterCompleteReceived event method added to
continue with client start up
- Client.GUI.cs checkForClient method renamed to clientExists and
transformed into a gettable bool for ease of use and continuity purposes
- Client.GUI.cs checkForSelfUpdate method reimplemented (as part of self
updating overhaul)
- Client.GUI.cs OnSelfUpdateCheckCompleted event method implemented to
trigger self update or proceed with client startup
- Client.GUI.cs updateClient method removed in favor of calling
UpdateHandler.Instance.Start() in line
- Client.GUI.cs added bool isMaintenance to OnArgumentsReceived event
method to implement maintenance mode
- Client.Network.ServerManager.cs Connection exception now allows user
to change IP/PORT in the case the Client was deployed with invalid
IP/PORT configuration
- Client.Network.ServerPackets.cs packet 0x000B changes from
SC_UpdateSelfUpdate to SC_SelfUpdateRequired
- Client.Network.ServerPackets.cs SC_UpdateSelfUpdate method replaced by
SC_SelfUpdateRequired
- Client.Network.ServerPackets.cs SC_ReceiveArguments method updates to
implement maintenance mode
- Client.Network.ServerPackets.cs CS_RegisterClient, CS_UnregisterClient
implemented to assist server in statistical recording
- Server.Functions.Database.cs updates to use new output system
- Server.Functions.IndexManager.cs Build method adjusted to use new
output system, lock and unlock it and slo trigger servers wait mode on
rebuilding
- Server.Functions.OPT.cs legacyList, deleteList renamed to Index for
continuity
- Server.Functions.OPT.cs AlterLegacyIndex/AlterDeleteIndex methods
added to manipulate their respective lists
- Server.Functions.OPT.cs SaveIndex added to save changes made via
Alter*Index methods
- Server.Functions.Output.cs completely rewritten for cleanliness and
ease of use, WriteAndLock, WriteAndUnlock methods were added for the
ability to lock a set of messages while queuing incoming messages.
- Server.Functions.Statistics.cs added to record, manipulate and send
statistical information to the updates Server GUI
- Server.Functions.UpdateHandler.cs updated to use new output system
- Server.Functions.UpdateHandler.cs OnUserRequestSelfUpdate replaced by
OnUserRequestSelfUpdateRequired
- Server.Functions.UserHandler.cs updated to use new output system
- Server.GUI.cs major updates to accomodate conversion from Command Line
(console) to GUI (form)
- Server.Network.ClientManager.cs updates to use new output system
- Server.Network.ClientPackets.cs  packet 0x0002 CS_ReququestSelfUpdate
replaced by CS_RequestSelfUpdateRequired
- Server.Network.ClientPackets.cs CS_RegisterClient /
CS_UnregisterClient, US_RegisterUpdater, CS_UnregisterUpdater methods
added to assist in statistical recording
- Server.settingsGUI.cs added to allow direct manipulation of the
portal.opt from within the Server GUI
- Server.Structures.GridProperty.cs added to faciliate the display and
alterations of portal.opt settings via the settingsGUI
- Server.Structures.Message.cs added to assist in the functionality of
the new output system
- Server.Structures.User.cs added for future functionality
- Server.updatesViewerGUI.cs added to allow manipulation of the update
index, legacy.opt and delete.opt from within the Server GUI
- Updater received major overhaul to connect to the Portal Server and
request updates in a manner similar to the client updating process for
safety and stability

+++ Updater/Functions/UpdateHandler.cs
@@ -27,7 +27,12 @@ namespace Updater.Functions
             string launcherPath = string.Format(@"{0}\{1}", Directory.GetCurrentDirectory(), "Launcher.exe");
             string backupPath = string.Format(@"{0}\{1}", Directory.GetCurrentDirectory(), "Launcher.exe.old");
 
-            if (File.Exists(launcherPath)) { File.Move(launcherPath, backupPath); }
+            if (!File.Exists(backupPath))
+            {
+                if (File.Exists(launcherPath)) { File.Move(launcherPath, backupPath); }
+            }
+            else { File.Delete(launcherPath); }
+            
 
             using (FileStream fs = File.Create(launcherPath))
             {
@@ -49,6 +54,7 @@ namespace Updater.Functions
             else
             {
                 Console.WriteLine("\tLauncher update completed!");
+                File.Delete(backupPath);
                 Program.OnLauncherVerified();
             }
         }
